[CODE] Add generic ChromiumBasedBrowser class
